backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlmodel==0.0.14
pytest==7.4.3
pytest-asyncio==0.21.1
pyyaml==6.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

